fast api hackathon/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
pydantic-settings>=2.3,<3.0
email-validator>=2.2,<3.0
